backend/pyproject.toml
[project]
name = "airdrop-works-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "django>=5.0",
    "djangorestframework>=3.15",
    "djangorestframework-simplejwt>=5.3",
    "django-cors-headers>=4.3",
    "dj-database-url>=2.1",
    "psycopg[binary]>=3.1",
    "redis>=5.0",
    "celery>=5.3",
    "django-celery-beat>=2.6",
    "resend>=2.0",
    "gunicorn>=21.2",
    "sentry-sdk[django]>=1.40",
    "python-decouple>=3.8",
    "anthropic>=0.25.0",
    "siwe>=4.0",
    "web3>=6.0",
    "stripe>=8.0",
    "pillow>=10.0",
    "django-storages>=1.14",
    "boto3>=1.34",
    "neo4j>=6.1.0",
    "django-extensions>=3.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.7",
    "pytest-cov>=4.1",
    "factory-boy>=3.3",
    "faker>=24.0",
    "ruff>=0.3",
    "pre-commit>=3.6",
    "ipython>=8.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.local"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--tb=short -v"

[tool.coverage.run]
source = ["apps"]
omit = ["*/migrations/*", "*/tests/*"]
